Rename HeyOrca's key
HeyOrca was one of the first companies to have jobs posted to the
website. At that time it was called whalecompany, but since the list has
grown and other companies have not been given nicknames, it makes sense
to properly rename the key for HeyOrca.

This change replaces all occurrences of "whalecompany" with "heyorca".